### 2025-04-07 #8016 @editorialbot
[PRE REVIEW]: ChunkitApp 2.0, a speech segmentation app
**Submitting author:** @akonina (<a href="http://orcid.org/0000-0002-1215-1514">Alena Konina</a>)
**Repository:** https://github.com/annthehuman/ChunkitApp2.0
**Branch with paper.md** (empty if default branch):
**Version:** 2.0
