### 2025-04-08 #8016 @akonina
The statement of purpose and .bib file are located in the 'paper' folder in
the main branch.

пн, 7 апр. 2025 г. в 12:52, The Open Journals editorial robot <
***@***.***>:

> *License info*:
>
> 🟡 License found: Other (Check here <https://opensource.org/licenses> for
> OSI approval)
>
