### 2025-04-10 #8016 @akonina
@samhforbes I can see that my file is .rmd instead of .md, maybe that's the issue. I'll fix it
