### 2025-04-10 #8016 @samhforbes
Oh sorry I didn't see that, that will be the issue. Yes if in doubt you can compile the rmd to GitHub markdown I think?
