### 2025-04-10 b96cd6f
Add files via upload
Adding a .md file to comply with JOSS publication rules
