### 2025-04-10 #8016 @akonina
@samhforbes I changed the output to github markdown but it still gives an issue. Now I am out of my depth:) Is there something else I can try?
