### 2025-04-10 #8016 @samhforbes
OK it looks like it needs the right title and yaml in the markdown file.
https://joss.readthedocs.io/en/latest/paper.html#article-metadata has how it needs to be formatted.
