+++ paper.md
@@ -3,13 +3,10 @@ title: "ChunkitApp 2.0 a speech segmentation app"
 authors:
   - name: Anna Galkova
     affiliation: ''
-
   - name: Alena Konina
     affiliation: '1'
-    
   - name: Anna Mauranen
     affiliation: '1'
-
 affiliations:
   - index: 1
     name: University of Helsinki
