+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ChunkitApp 2.0 a speech segmentation app'
+title:'ChunkitApp 2.0 a speech segmentation app'
 authors:
   - name: Anna Galkova
     affiliation: ''
