### 2025-04-23 #8016 @samhforbes
Hi @bagustris, @sdruskat, @neillu23 would you be interested in reviewing this software for JOSS?
