### 2025-04-23 #8016 @akonina
@samhforbes I will change the README.txt accordingly by end of day today, thank you for pointing that out.
