### 2025-06-25 #8486 @wmvanvliet
-  [x] https://github.com/annthehuman/ChunkitApp2.0/issues/3

  - [x] https://github.com/annthehuman/ChunkitApp2.0/issues/4

  - [x] https://github.com/annthehuman/ChunkitApp2.0/issues/5
  - [x] https://github.com/annthehuman/ChunkitApp2.0/issues/9
  - [x] https://github.com/annthehuman/ChunkitApp2.0/issues/10
  - [ ] https://github.com/annthehuman/ChunkitApp2.0/issues/6
  - [ ]  https://github.com/annthehuman/ChunkitApp2.0/issues/7
  - [ ] https://github.com/annthehuman/ChunkitApp2.0/issues/8
