### 2025-06-25 annthehuman/ChunkitApp2.0#3 @wmvanvliet
[JOSS review] Conflicting license information.
The README states: This project is distributed under CC-BY-NC-SA license.
however, the LICENSE.txt contains the Academic Free License v. 3.0.

Which license is the intended one?
