### 2025-06-25 annthehuman/ChunkitApp2.0#4 @wmvanvliet
[JOSS Review] Installation instructions incomplete
I could follow the installation instructions to get the webapp running on my local machine. Browsing to 127.0.0.1 showed a login box. Signing up as a new user did not work, and is there some sort of default admin account?

This is the error log when trying to sign up:

```
proxy-1    | 172.18.0.1 - - [25/Jun/2025:21:21:26 +0000] "POST /auth/users/ HTTP/1.1" 500 16217 "http://127.0.0.1/signup/" "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:140.0) Gecko/20100101 Firefox/140.0" "-"
app-1      |   File "/py/lib/python3.9/site-packages/django/core/handlers/exception.py", line 47, in inner
app-1      |     response = get_response(request)
app-1      |   File "/py/lib/python3.9/site-packages/django/core/handlers/base.py", line 179, in _get_response
app-1      |     response = wrapped_callback(request, *callback_args, **callback_kwargs)
app-1      |   File "/py/lib/python3.9/site-packages/django/views/decorators/csrf.py", line 54, in wrapped_view
app-1      |     return view_func(*args, **kwargs)
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/viewsets.py", line 125, in view
app-1      |     return self.dispatch(request, *args, **kwargs)
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/views.py", line 509, in dispatch
app-1      |     response = self.handle_exception(exc)
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/views.py", line 469, in handle_exception
app-1      |     self.raise_uncaught_exception(exc)
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/views.py", line 480, in raise_uncaught_exception
app-1      |     raise exc
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/views.py", line 506, in dispatch
app-1      |     response = handler(request, *args, **kwargs)
app-1      |   File "/py/lib/python3.9/site-packages/rest_framework/mixins.py", line 19, in create
app-1      |     self.perform_create(serializer)
app-1      |   File "/py/lib/python3.9/site-packages/djoser/views.py", line 144, in perform_create
app-1      |     settings.EMAIL.activation(self.request, context).send(to)
app-1      |   File "/py/lib/python3.9/site-packages/templated_mail/mail.py", line 78, in send
app-1      |     super(BaseEmailMessage, self).send(*args, **kwargs)
app-1      |   File "/py/lib/python3.9/site-packages/django/core/mail/message.py", line 284, in send
app-1      |     return self.get_connection(fail_silently).send_messages([self])
app-1      |   File "/py/lib/python3.9/site-packages/django/core/mail/backends/smtp.py", line 102, in send_messages
app-1      |     new_conn_created = self.open()
app-1      |   File "/py/lib/python3.9/site-packages/django/core/mail/backends/smtp.py", line 69, in open
app-1      |     self.connection.login(self.username, self.password)
app-1      |   File "/usr/local/lib/python3.9/smtplib.py", line 750, in login
app-1      |     raise last_exception
app-1      |   File "/usr/local/lib/python3.9/smtplib.py", line 739, in login
app-1      |     (code, resp) = self.auth(
app-1      |   File "/usr/local/lib/python3.9/smtplib.py", line 662, in auth
app-1      |     raise SMTPAuthenticationError(code, resp)
app-1      | smtplib.SMTPAuthenticationError: (535, b'5.7.8 Username and Password not accepted. For more information, go to\n5.7.8  https://support.google.com/mail/?p=BadCredentials 38308e7fff4ca-32b980dd1b1sm21387231fa.96 - gsmtp')
```
