### 2025-06-26 annthehuman/ChunkitApp2.0#5 @wmvanvliet
[JOSS review] No "statement of need" in the README
It would be great if the README would begin with an introduction section stating what ChunkitApp does and include a screenshot of it in action (as example usage).
