### 2025-06-26 annthehuman/ChunkitApp2.0#6 @wmvanvliet
[JOSS review] No functional documentation
Currently, no documentation other than some installation instructions are provided. To comply with JOSS requirements, a software package needs thorough documentation of its features and how to use them.
