### 2025-06-26 annthehuman/ChunkitApp2.0#7 @wmvanvliet
[JOSS review] No automated tests
There are currently no automated tests to verify the correct functioning of the software.
