### 2025-06-26 annthehuman/ChunkitApp2.0#8 @wmvanvliet
[JOSS review] No community guidelines
There are some community guidelines in the paper, but they are missing from the README (a separate CONTRIBUTING file is also an option). We need clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. This includes how to setup the software locally in "development" mode.
