### 2025-06-26 annthehuman/ChunkitApp2.0#9 @wmvanvliet
[JOSS review] No description of the software stack
With docker-compose, ChunkitApp was easy to build and start. However, it would still be nice to have a description of the exact software stack used so the user knows what they are running on their server.
