### 2025-06-26 annthehuman/ChunkitApp2.0#10 @wmvanvliet
[JOSS review] No example usage in the README
It would be great if the README could show off the package a little (screenshots, perhaps even a little animation?), so user could see at a glance what it is capable of and whether this would suit their needs.
