### 2025-06-27 #8486 @akonina
@wmvanvliet thank you for raising the issue! @annthehuman could you please help with creating a new user?
