+++ DOCUMENTATION.md
@@ -1,0 +1,587 @@
+# ChunkitApp 2.0 Documentation
+
+## 1. Overview (High-Level Summary)
+
+### Purpose
+ChunkitApp 2.0 is a web-based data collection platform for speech segmentation experiments. The app allows researchers to design and run experiments where participants segment orthographic transcripts of speech extracts by tapping interactive symbols between words. It aggregates data across participants and extracts, and can run Monte Carlo simulations to analyze segmentation behavior.
+
+### Target Users
+- Linguists and speech researchers
+- Academic researchers studying language processing and speech segmentation
+- Research groups investigating chunking in language perception
+
+### Key Features
+- **Experiment Design**: Create custom speech segmentation experiments with audio files and transcripts
+- **Interactive Transcripts**: Participants tap symbols between words to mark segment boundaries
+- **Multi-modal Data Collection**: Background questionnaires, feedback forms, and comprehension questions
+- **English Proficiency Testing**: Built-in imitation tasks for assessing participants' English proficiency
+- **Data Aggregation**: Automatic aggregation of segmentation data across participants
+- **Statistical Analysis**: Monte Carlo simulations and permutation tests for significance testing
+- **Result Export**: CSV export of raw and processed data
+- **User Authentication**: JWT-based authentication system with email activation
+- **Draft Management**: Save and load experiment drafts
+- **Audio File Management**: Support for ZIP archives containing audio files
+- **Prolific Integration**: Built-in support for Prolific participant recruitment
+
+## 2. Tech Stack
+
+### Backend
+- **Django**: 3.1.3 (Python web framework)
+- **Django REST Framework**: 3.13.1 (API development)
+- **djoser**: 2.1.0 (Authentication endpoints)
+- **django-cors-headers**: 3.8.0 (CORS handling)
+- **djangorestframework-simplejwt**: 4.8.0 (JWT authentication)
+
+### Frontend
+- **React**: 17.0.2 (JavaScript library)
+- **React Router DOM**: 5.2.0 (Client-side routing)
+- **Material-UI**: 5.5.2 (@mui/material, @mui/icons-material)
+- **Bootstrap**: 4.6.0 (CSS framework)
+- **Styled Components**: 5.3.0 (CSS-in-JS)
+- **Draft.js**: 0.11.7 (Rich text editor)
+- **Papa Parse**: 5.3.1 (CSV parsing)
+
+### Data Processing & Analysis
+- **NumPy**: 1.21.4 (Numerical computing)
+- **Pandas**: 1.3.4 (Data manipulation)
+- **SciPy** (Statistical functions)
+- **python-Levenshtein**: 0.12.2 (String distance calculations)
+- **Matplotlib**, **Seaborn**, **scikit-image** (Data visualization and analysis)
+
+### Database
+- **SQLite3** (Development and production database)
+
+### Deployment & Infrastructure
+- **Docker**: 5.0.2 (Containerization)
+- **Docker Compose**: 1.29.2 (Multi-container orchestration)
+- **Nginx** (Reverse proxy and static file serving)
+- **uWSGI**: >=2.0.19.1,<2.1 (WSGI server)
+- **Certbot** (SSL certificate management)
+
+### Build Tools
+- **Webpack**: 5.47.1 (Module bundler)
+- **Babel**: 7.14.8 (JavaScript transpiler)
+
+## 3. Setup Guide
+
+### Prerequisites
+- Docker and Docker Compose installed
+- Git installed
+- Minimum 4GB RAM
+- **MISSING**: Specific Python version requirements beyond Python 3.9 (used in Dockerfile)
+
+### Installation
+
+1. **Clone the repository:**
+   ```bash
+   git clone https://github.com/annthehuman/ChunkitApp2.0.git
+   cd ChunkitApp2.0
+   ```
+
[…]
