### 2025-08-13 20fe47a
upgrading the license type to Acamedic Free License v. 3.0
