+++ README.md
@@ -33,4 +33,4 @@ You can access your application in a browser at 127.0.0.1 on local machine or at
 
 If you have any questions about the software part of the application, please contact me on Github. For any other questions, please contact Alena Konina on Github or at alena.konina@helsinki.fi.
 
-## This project is distributed under CC-BY-NC-SA license.
+## This project is distributed under Academic Free License v. 3.0 license.
