+++ chunkitapp/demo.py
@@ -1,0 +1,131 @@
+from django.conf import settings
+from .models import draft_data, experiment_links
+import os
+import shutil
+import pandas as pd
+
+def create_demo_experiment():
+    """
+    Create a demo experiment using pre-loaded audio files and transcripts
+    from the data/demo folder.
+    
+    This function:
+    1. Creates demo copies of audio and transcript files in the media folder
+    2. Creates a draft_data entry for the demo experiment
+    3. Creates an experiment_links entry for the demo experiment
+    4. Returns the experiment name and access token
+    """
+    
+    demo_experiment_name = 'Demo_Experiment'
+    demo_folder = os.path.join(settings.BASE_DIR, 'data', 'demo')
+    media_folder = settings.MEDIA_ROOT
+    
+    # Paths to demo files
+    demo_audio_zip = os.path.join(demo_folder, 'demo_extracts.zip')
+    demo_transcripts = os.path.join(demo_folder, 'transcripts_demo', 'transcripts_demo.xlsx')
+    
+    # Verify demo files exist
+    if not os.path.exists(demo_audio_zip):
+        raise FileNotFoundError(f"Demo audio file not found: {demo_audio_zip}")
+    if not os.path.exists(demo_transcripts):
+        raise FileNotFoundError(f"Demo transcripts file not found: {demo_transcripts}")
+    
+    # Copy demo files to media folder with unique names
+    demo_audio_dest = os.path.join(media_folder, 'demo_audio.zip')
+    demo_transcripts_dest = os.path.join(media_folder, 'demo_transcripts.xlsx')
+    
+    # Copy files
+    shutil.copy2(demo_audio_zip, demo_audio_dest)
+    shutil.copy2(demo_transcripts, demo_transcripts_dest)
+    
+    # Remove any existing demo experiment
+    try:
+        existing = draft_data.objects.filter(nameExperementForParticipants=demo_experiment_name)
+        existing.delete()
+        existing_links = experiment_links.objects.filter(experiment_link__contains=demo_experiment_name)
+        existing_links.delete()
+    except:
+        pass
+    
+    # Read transcript data from Excel file
+    df = pd.read_excel(demo_transcripts, 
+                       converters={'Audio name': str, 'Transcript': str,
+                                   'Question': str, 'Answer1': str,
+                                   'Answer2': str}).fillna(0)
+    
+    transcripts_data = []
+    for row in range(df.shape[0]):
+        table_row = []
+        for col in range(df.shape[1]):
+            table_row.append(df.iat[row, col])
+        transcripts_data.append(table_row)
+    
+    # Get list of audio files
+    audio_files = ['Demo_Experiment/sonnet_01_01.mp3', 
+                   'Demo_Experiment/sonnet_01_02.mp3', 
+                   'Demo_Experiment/sonnet_01_03.mp3']
+    
+    # Create demo experiment draft
+    demo_draft = draft_data.objects.create(
+        accessToken='demo_token_001',
+        nameExperement='Demo Experiment',
+        sessionTime=90,
+        nameExperementForParticipants=demo_experiment_name,
+        ImitationTask=False,
+        UseQuestions=True,
+        UseProlific=False,
+        shuffleExtracts='false',
+        shuffleExtractsPractice='false',
+        helloEditor='<p>Welcome to ChunkitApp Demo!</p>',
+        consentEditor='<p>By continuing, you agree to participate in this demo.</p>',
+        outlineEditor='<p>In this demo, you will listen to speech extracts and mark segment boundaries.</p>',
+        backgroundExample='{}',
+        backgroundAddQ='{}',
+        feedbackExample='{}',
+        feedbackAddQ='{}',
[…]
