+++ chunkitapp_project/settings.py
@@ -83,11 +83,17 @@ REST_FRAMEWORK = {
 }
 
 #setting up email server
-EMAIL_HOST = 'smtp.gmail.com'
-EMAIL_HOST_USER = 'info.chunkitapp@gmail.com'
-EMAIL_HOST_PASSWORD = 'ppppppp'
-EMAIL_PORT = 587
-EMAIL_USE_TLS = True
+# For development, use console backend to avoid SMTP authentication errors
+# For production, configure proper email credentials
+if DEBUG:
+    EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
+else:
+    EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
+    EMAIL_HOST = 'smtp.gmail.com'
+    EMAIL_HOST_USER = os.environ.get('EMAIL_HOST_USER', 'info.chunkitapp@gmail.com')
+    EMAIL_HOST_PASSWORD = os.environ.get('EMAIL_HOST_PASSWORD', 'ppppppp')
+    EMAIL_PORT = 587
+    EMAIL_USE_TLS = True
 
 # DOMAIN = ('51.210.4.176')
 
