+++ scripts/run.sh
@@ -5,4 +5,7 @@ set -e
 python manage.py collectstatic --noinput
 python manage.py migrate
 
+# Create symlink for media files
+ln -sf /vol/web/media /vol/web/static/media
+
 uwsgi --socket :9000 --workers 4 --master --enable-threads --module chunkitapp_project.wsgi
