+++ chunkitapp_project/settings.py
@@ -12,6 +12,7 @@ https://docs.djangoproject.com/en/3.2/ref/settings/
 from datetime import timedelta
 from pathlib import Path
 import os
+import sys
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
@@ -179,6 +180,10 @@ DATABASES = {
         #'NAME': './db.sqlite3',
     }
 }
+
+# Use local SQLite DB for tests or when explicitly requested
+if 'test' in sys.argv or os.environ.get('USE_LOCAL_DB') == '1':
+    DATABASES['default']['NAME'] = os.path.join(BASE_DIR, 'db.sqlite3')
 
 
 
