+++ chunkitapp/views.py
@@ -1,4 +1,3 @@
-from distutils.log import error
 from django.shortcuts import redirect
 from django.http import HttpResponse
 from .models import background, feedback, test, experiment_links
@@ -312,6 +311,154 @@ def load_draft(request):
         return HttpResponse(json.dumps(draft_data_values))
 
 
+def _parse_dt_maybe(value):
+    """Parse datetime stored as text in DB, return None if not parseable."""
+    try:
+        return datetime.datetime.strptime(value, '%Y-%m-%d %H:%M:%S.%f')
+    except Exception:
+        return None
+
+
+def _get_session_time_minutes(experiment_name):
+    """Return session duration in minutes for experiment (defaults to 90)."""
+    try:
+        return int(list(draft_data.objects.filter(
+            nameExperementForParticipants=experiment_name
+        ).values_list('sessionTime', flat=True))[0])
+    except Exception:
+        return 90
+
+
+def _get_last_link_row(experiment_name):
+    rows = list(experiment_links.objects.filter(
+        experiment_link__iregex=rf'experiment/{experiment_name}.*'
+    ).values_list())
+    return rows[-1] if rows else None
+
+
+def _now():
+    return datetime.datetime.now()
+
+
+def _basename_no_ext(path):
+    try:
+        base = os.path.basename(path)
+        return os.path.splitext(base)[0]
+    except Exception:
+        return ''
+
+
+def _literal_eval_safely(value, default):
+    try:
+        return ast.literal_eval(value)
+    except Exception:
+        return default
+
+
+def experiment_status(request):
+    """
+    Return JSON status for experiment timing based on start time and sessionTime.
+    GET params: name=<experiment_name>
+    { over: bool, start: str|None, end: str|None, remaining_seconds: int|None }
+    """
+    if request.method == 'GET':
+        experiment_name = request.GET.get('name', '')
+        link_row = _get_last_link_row(experiment_name)
+        if not link_row:
+            return HttpResponse(json.dumps({
+                'over': False,
+                'start': None,
+                'end': None,
+                'remaining_seconds': None
+            }))
+
+        # experiment_links fields order (inspect usages in code):
+        # (..., experiment_start_time index likely 4, experiment_stopped index likely -2)
+        start_raw = link_row[4]
+        start_dt = _parse_dt_maybe(start_raw) if start_raw and start_raw != 'nothing' else None
+        if not start_dt:
+            return HttpResponse(json.dumps({
+                'over': False,
+                'start': None,
+                'end': None,
+                'remaining_seconds': None
+            }))
+
+        minutes = _get_session_time_minutes(experiment_name)
+        end_dt = start_dt + datetime.timedelta(minutes=minutes)
+        now = _now()
+        over = now >= end_dt
+        remaining = int((end_dt - now).total_seconds()) if not over else 0
+        return HttpResponse(json.dumps({
+            'over': over,
+            'start': str(start_dt),
+            'end': str(end_dt),
+            'remaining_seconds': remaining
+        }))
+
+
+def validate_uploads(request):
+    """
+    Validate that number of audio extracts equals number of transcript rows and
+    that transcript names correspond to audio basenames.
+    GET params: name=<experiment_name>
+    { ok: bool, warning: str, audio_count: int, transcript_count: int }
+    """
+    if request.method == 'GET':
+        experiment_name = request.GET.get('name', '')
+        qs = list(draft_data.objects.filter(
+            nameExperementForParticipants=experiment_name
+        ).values_list())
+        if not qs:
+            return HttpResponse(json.dumps({
+                'ok': False,
+                'warning': 'Experiment not found',
+                'audio_count': 0,
+                'transcript_count': 0
+            }), status=404)
+
[…]
