+++ chunkitapp/tests.py
@@ -1,3 +1,610 @@
-from django.test import TestCase
+from django.test import TestCase, Client
+from django.contrib.auth.models import User
+from django.urls import reverse
+from rest_framework.authtoken.models import Token
+from rest_framework.test import APITestCase, APIClient
+from .models import (
+    draft_data, 
+    experiment_links, 
+    background, 
+    feedback, 
+    test, 
+    sentence
+)
+import json
+import tempfile
+import os
+from django.core.files.uploadedfile import SimpleUploadedFile
 
-# Create your tests here.
+
+class UserCreationTests(APITestCase):
+    """Test user creation and authentication"""
+    
+    def setUp(self):
+        self.client = APIClient()
+        self.user_data = {
+            'username': 'testuser',
+            'email': 'testuser@example.com',
+            'password': 'TestPassword123!',
+        }
+    
+    def test_create_user(self):
+        """Test that a user can be created"""
+        user = User.objects.create_user(
+            username=self.user_data['username'],
+            email=self.user_data['email'],
+            password=self.user_data['password']
+        )
+        self.assertEqual(user.username, 'testuser')
+        self.assertEqual(user.email, 'testuser@example.com')
+        self.assertTrue(user.check_password('TestPassword123!'))
+    
+    def test_user_authentication(self):
+        """Test that a user can authenticate and get a token"""
+        user = User.objects.create_user(
+            username=self.user_data['username'],
+            email=self.user_data['email'],
+            password=self.user_data['password']
+        )
+        token, created = Token.objects.get_or_create(user=user)
+        self.assertIsNotNone(token)
+        self.assertTrue(created)
+    
+    def test_duplicate_username_fails(self):
+        """Test that creating a user with duplicate username fails"""
+        User.objects.create_user(
+            username=self.user_data['username'],
+            email=self.user_data['email'],
+            password=self.user_data['password']
+        )
+        with self.assertRaises(Exception):
+            User.objects.create_user(
+                username=self.user_data['username'],
+                email='another@example.com',
+                password='AnotherPassword123!'
+            )
+
+
+class ExperimentCreationTests(TestCase):
+    """Test experiment creation and management"""
+    
+    def setUp(self):
+        self.client = Client()
+        self.user = User.objects.create_user(
+            username='testuser',
+            email='test@example.com',
+            password='testpass123'
+        )
+        self.token = Token.objects.create(user=self.user)
+    
+    def test_create_experiment_link(self):
+        """Test creating an experiment link"""
+        experiment_link = experiment_links.objects.create(
+            experiment_link='experiment/test_experiment',
+            accessToken='test_token_123',
+            experiment_stopped='False'
+        )
+        self.assertEqual(experiment_link.experiment_link, 'experiment/test_experiment')
+        self.assertEqual(experiment_link.accessToken, 'test_token_123')
+        self.assertEqual(experiment_link.experiment_stopped, 'False')
+    
+    def test_create_draft_experiment(self):
+        """Test creating a draft experiment"""
+        draft = draft_data.objects.create(
+            accessToken='test_token_123',
+            nameExperement='Test Experiment',
+            nameExperementForParticipants='test_experiment',
+            shuffleExtracts='True',
+            shuffleExtractsPractice='True',
+            ImitationTask=False,
+            UseQuestions=True,
+            UseProlific=False,
+            linkToProlific='',
+            helloEditor='Hello!',
+            consentEditor='Please consent',
+            outlineEditor='Outline here',
[…]
