+++ README.md
@@ -31,6 +31,26 @@ to run it.
 
 You can access your application in a browser at 127.0.0.1 on local machine or at the ip adress of your server.
 
+## Login and Test Account
+
+After installation, you can log in using the pre-configured test account:
+
+```
+Email/Username: test@chunkit.app
+Password:       test1234
+```
+
+**Note about new user registration:** New user signups require email activation. In development mode (DEBUG=1), activation emails are printed to the Docker logs instead of being sent. To activate a new account:
+
+1. Sign up as a new user
+2. Check the Docker logs for the activation link:
+   ```sh
+   docker-compose logs app
+   ```
+3. Copy and paste the activation URL into your browser
+
+Alternatively, just use the test account above for immediate access.
+
 ## Demo Feature
 
 ChunkitApp includes a **Demo Mode** that allows visitors to try the application without creating an account. The demo loads pre-configured experiment files from the `data/demo` folder and provides a quick way to explore the app's functionality.
@@ -62,19 +82,22 @@ To change the demo content:
 3. Delete the `data/demo/demo_extracts.zip` file (it will be regenerated automatically)
 4. Restart the application
 
-## Test Account
+## Troubleshooting
 
-For development and testing, a universal test account is available:
+### Reset Test Account Password
 
-```
-Email/Username: test@chunkit.app
-Password:       test1234
-```
-
-This account is pre-activated and ready to use. To reset the password or recreate the account, run:
+If you need to reset the test account password (test@chunkit.app), run:
 
 ```sh
 docker exec chunkitapp20-app-1 python -c "from django.contrib.auth.models import User; u = User.objects.get(username='test@chunkit.app'); u.set_password('test1234'); u.save(); print('Password reset!')"
+```
+
+### Email Activation Issues
+
+If you signed up as a new user and didn't receive an activation email, remember that in development mode emails are printed to the Docker logs. Check them with:
+
+```sh
+docker-compose logs app | grep -A 20 "activation"
 ```
 
 If you have any questions about the software part of the application, please contact me on Github. For any other questions, please contact Alena Konina on Github or at alena.konina@helsinki.fi.
