+++ DOCUMENTATION.md
@@ -6,21 +6,15 @@
 ChunkitApp 2.0 is a web-based data collection platform for speech segmentation experiments. The app allows researchers to design and run experiments where participants segment orthographic transcripts of speech extracts by tapping interactive symbols between words. It aggregates data across participants and extracts, and can run Monte Carlo simulations to analyze segmentation behavior.
 
 ### Target Users
-- Linguists and speech researchers
-- Academic researchers studying language processing and speech segmentation
-- Research groups investigating chunking in language perception
+- Researchers studying language processing, speech segmentation, and chunking
 
 ### Key Features
-- **Experiment Design**: Create custom speech segmentation experiments with audio files and transcripts
-- **Interactive Transcripts**: Participants tap symbols between words to mark segment boundaries
-- **Multi-modal Data Collection**: Background questionnaires, feedback forms, and comprehension questions
-- **English Proficiency Testing**: Built-in imitation tasks for assessing participants' English proficiency
+- **Experiment Design**: Create custom speech segmentation experiments with audio files and transcripts. The app shows the transcripts as the speech extract plays. Participants in an experiment can mark boundaries between segments by tapping an interactive symbol between words and take them away by tapping it again. The experimental design is flexible: researchers can choose which experiment features to include.
+- **Flexible Data Collection**: Create your own or choose sample background questionnaires, feedback forms, and comprehension questions around your segmentation task.
+- **English Proficiency Testing**: Optional built-in imitation tasks for assessing participants' English proficiency.
 - **Data Aggregation**: Automatic aggregation of segmentation data across participants
 - **Statistical Analysis**: Monte Carlo simulations and permutation tests for significance testing
 - **Result Export**: CSV export of raw and processed data
-- **User Authentication**: JWT-based authentication system with email activation
-- **Draft Management**: Save and load experiment drafts
-- **Audio File Management**: Support for ZIP archives containing audio files
 - **Prolific Integration**: Built-in support for Prolific participant recruitment
 
 ## 2. Tech Stack
@@ -107,6 +101,13 @@ ChunkitApp 2.0 is a web-based data collection platform for speech segmentation e
    - Local: http://127.0.0.1
    - Server: http://your_server_ip
 
+6. **Login with test account:**
+   ```
+   Email/Username: test@chunkit.app
+   Password:       test1234
+   ```
+or **Access the Demo**
+
 ### Development Setup
 
 For local development without Docker:
@@ -157,10 +158,13 @@ For local development without Docker:
 
 ### Email Configuration
 
+The application uses email for user account activation. Email behavior differs between development and production:
+
 **Development Mode (DEBUG=1):**
 - Uses `django.core.mail.backends.console.EmailBackend`
 - Emails are printed to console/Docker logs instead of being sent
 - No SMTP credentials required
+- View activation emails with: `docker-compose logs app | grep -A 20 "activation"`
 
 **Production Mode (DEBUG=0):**
 - Uses `django.core.mail.backends.smtp.EmailBackend`
@@ -176,6 +180,8 @@ For local development without Docker:
 export EMAIL_HOST_USER=your-email@gmail.com
 export EMAIL_HOST_PASSWORD=your-app-password  # Use Gmail App Password, not regular password
 ```
+
+**Note:** For detailed information on user registration and email activation, see the "User Registration and Email Activation" section in the Testing chapter.
 
[…]
