### 2025-10-20 annthehuman/ChunkitApp2.0#7 @akonina
@wmvanvliet added them to the best of my ability
