### 2025-10-20 annthehuman/ChunkitApp2.0#6 @akonina
@wmvanvliet clarified instructions for running the app locally, added a demo and a universal test account for users of various level of coding skills
