### 2025-10-20 68f8710
add demo video placeholder
