### 2025-10-20 annthehuman/ChunkitApp2.0#4 @wmvanvliet
`docker compose build` works, but `docker compose up` now throws up the following errors:

```
> docker compose up
time="2025-10-21T00:05:43+03:00" level=warning msg="The \"d38\" variable is not set. Defaulting to a blank string."
time="2025-10-21T00:05:43+03:00" level=warning msg="The \"d38\" variable is not set. Defaulting to a blank string."
time="2025-10-21T00:05:43+03:00" level=warning msg="The \"d38\" variable is not set. Defaulting to a blank string."
time="2025-10-21T00:05:43+03:00" level=warning msg="The \"d38\" variable is not set. Defaulting to a blank string."
time="2025-10-21T00:05:43+03:00" level=warning msg="C:\\Users\\wmvan\\projects\\ChunkitApp2.0\\docker-compose.yml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion"
[+] Running 4/4
 ✔ Container chunkitapp20-app-1      Recreated                                                                                                                  21.4s
 ✔ Container chunkitapp20-db-1       Recreated                                                                                                                   0.1s
 ✔ Container chunkitapp20-proxy-1    Recreated                                                                                                                  11.2s
 ✔ Container chunkitapp20-certbot-1  Recreated                                                                                                                   0.3s
Attaching to app-1, certbot-1, db-1, proxy-1
db-1       | exec /scripts/run.sh: no such file or directory
db-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1 exited with code 255
certbot-1  | Pooping...
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1 exited with code 127
certbot-1 exited with code 0
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1 exited with code 255
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1 exited with code 127
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1 exited with code 255
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
app-1      | exec /scripts/run.sh: no such file or directory
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1 exited with code 127
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
proxy-1    | /docker-entrypoint.sh: 47: exec: /run.sh: not found
```
