### 2025-10-20 dbe285e
add a video demo to README.md
