### 2025-10-21 34fb62c
add a fix to enforce Unix line index for Windows users
