### 2025-10-21 #8486 @wmvanvliet
Great work @akonina, you indeed addressed many of my comments. Unfortunately, I still cannot get the software to run properly 🥲 see issue: https://github.com/annthehuman/ChunkitApp2.0/issues/11

The demo works great though!
