### 2025-10-26 abf5afb
add fix for test account and user activation for Windows users
