### 2025-10-26 b0e6abc
fix errors with activation and demo run
