+++ frontend/src/components/logging-in/signup/signup.js
@@ -19,7 +19,9 @@ class Signup extends Component{
             emailSended: false,
             showPassword: false,
             passValid: true,
-            errorSingup: []
+            errorSingup: [],
+            activationStatus: null,
+            errorMessage: null
         };
 
         this.handleChange = this.handleChange.bind(this);
@@ -27,6 +29,7 @@ class Signup extends Component{
         this.handleClickShowPassword = this.handleClickShowPassword.bind(this);
         this.handleClickShowConfirmPassword = this.handleClickShowConfirmPassword.bind(this);
         this.handleMouseDownPassword = this.handleMouseDownPassword.bind(this);
+        this.handleAutoActivate = this.handleAutoActivate.bind(this);
     }
 
     handleChange(event) {
@@ -74,6 +77,46 @@ class Signup extends Component{
             }).catch((data) => {
             console.log(`Try again! Error: ${data}`)
             });
+    }
+
+    handleAutoActivate() {
+        // For development mode - try to auto-activate the user
+        this.setState({ activationStatus: 'activating' });
+        
+        // Try to activate the user by making a request to the auto-activation endpoint
+        // This is a development-only feature
+        fetch('/auto_activate/', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                email: this.state.email,
+                password: this.state.password
+            })
+        })
+        .then(response => response.json())
+        .then(data => {
+            if (data.success) {
+                this.setState({ 
+                    activationStatus: 'success',
+                    emailSended: false 
+                });
+                // Redirect to login after successful activation
+                setTimeout(() => {
+                    this.props.history.push('/login/');
+                }, 2000);
+            } else {
+                this.setState({ 
+                    activationStatus: 'error',
+                    errorMessage: data.error || 'Activation failed'
+                });
+            }
+        })
+        .catch(error => {
+            console.error('Auto-activation error:', error);
+            this.setState({ activationStatus: 'error' });
+        });
     }
 
     render() {
@@ -157,11 +200,61 @@ class Signup extends Component{
             </>
             :
             <>
-            <Typography>Thank you for registration!<br/>
-            Please сheck your email to confirm your account.</Typography>
-            <Link to='/login/'>
-            <CustomButton theme='blue' text='Log in'/>
-            </Link>
+            {this.state.activationStatus === 'success' ? (
+                <>
+                <Typography color="success.main">Account activated successfully!<br/>
+                Redirecting to login...</Typography>
+                </>
+            ) : this.state.activationStatus === 'error' ? (
+                <>
+                <Typography color="error.main">Auto-activation failed.<br/>
+                {this.state.errorMessage && <span>{this.state.errorMessage}</span>}
+                Please check the console logs for the activation link.</Typography>
+                <Typography variant="body2" sx={{ mt: 2, p: 2, backgroundColor: '#f5f5f5', borderRadius: 1, fontSize: '0.875rem' }}>
+                    <strong>Development Mode:</strong><br/>
+                    Check the console logs for the activation link:<br/>
+                    <code>docker-compose logs app | grep -A 20 "activation"</code>
+                </Typography>
+                <Stack direction="row" spacing={2} sx={{ mt: 2 }}>
+                    <Link to='/login/'>
+                        <CustomButton theme='blue' text='Log in'/>
+                    </Link>
+                    <CustomButton 
+                        theme='green' 
+                        text='Try Auto-activate Again' 
+                        onClick={this.handleAutoActivate}
+                    />
+                </Stack>
+                </>
+            ) : this.state.activationStatus === 'activating' ? (
+                <>
+                <Typography>Activating your account...</Typography>
+                </>
+            ) : (
[…]
