### 2025-10-26 3654b14
add a Windows fix for correct access to application
