+++ scripts/run.sh
@@ -5,6 +5,9 @@ set -e
 python manage.py collectstatic --noinput
 python manage.py migrate
 
+# Create test user account automatically
+python manage.py create_test_user
+
 # Create symlink for media files
 ln -sf /vol/web/media /vol/web/static/media
 
