### 2026-01-11 #8486 @wmvanvliet
@akonina My main issue is still the documentation. It is now very extensive but does not provide the information I actually need to properly set up the system. Before I get into more detail, may I ask if this is AI generated? I ask, because if it is, then we can more easily make sweeping changes to it. If it was painstakingly hard written we should of course strive to preserve it mas much as we can.
