### 2026-01-11 annthehuman/ChunkitApp2.0#6 @wmvanvliet
Sorry, but the instructions you added to the README are not helpful to me at all. They are very thorough, but cover topics that are not useful to me as someone looking to get Chunkit up and running in a smooth manner. The strange focus on explaining how the code works internally at such a level of detail makes me think this is AI generated. If so, please tell us so we take this into account when proposing changes (e.g. I would not lightly propose cutting or re-writing entire sections if this was all hand crafted).
