### 2026-01-15 annthehuman/ChunkitApp2.0#6 @akonina
@wmvanvliet the documentation was AI-generated, you are correct (I have very little experience with it). I could make it more concise and user-friendly with actual human feedback myself but would really appreciate any advice from you as well.
