### 2026-01-16 annthehuman/ChunkitApp2.0#6 @wmvanvliet
Here are some suggestions:

- [x] Split off sections 2, 4, 5, 6, 7 into a new document: `docs/architecture.md`
- [x] Get rid of section 8 (testing)
- [x] Merge section 9 (Deployment) with section 3 (Setup guide)
- [x] Split off section 10 into a new document: `docs/troubleshooting.md`
- [x] Split off section 11 into a new document: `CONTRIBUTING.md`
- [x] Get rid of duplication of windows troubleshooting points, make sure it is only in `docs/troubleshooting.md`

Let's see where we are then.
