### 2026-01-27 21d4345
shortening READ.me for a more user-friendly experience, add troubleshooting.md as architecture.md as separate files in the docs folder
