+++ docs/architecture.md
@@ -1,0 +1,419 @@
+# Architecture Documentation
+
+## Tech Stack
+
+### Backend
+- **Django**: 3.1.3 (Python web framework)
+- **Django REST Framework**: 3.13.1 (API development)
+- **djoser**: 2.1.0 (Authentication endpoints)
+- **django-cors-headers**: 3.8.0 (CORS handling)
+- **djangorestframework-simplejwt**: 4.8.0 (JWT authentication)
+
+### Frontend
+- **React**: 17.0.2 (JavaScript library)
+- **React Router DOM**: 5.2.0 (Client-side routing)
+- **Material-UI**: 5.5.2 (@mui/material, @mui/icons-material)
+- **Bootstrap**: 4.6.0 (CSS framework)
+- **Styled Components**: 5.3.0 (CSS-in-JS)
+- **Draft.js**: 0.11.7 (Rich text editor)
+- **Papa Parse**: 5.3.1 (CSV parsing)
+
+### Data Processing & Analysis
+- **NumPy**: 1.21.4 (Numerical computing)
+- **Pandas**: 1.3.4 (Data manipulation)
+- **SciPy** (Statistical functions)
+- **python-Levenshtein**: 0.12.2 (String distance calculations)
+- **Matplotlib**, **Seaborn**, **scikit-image** (Data visualization and analysis)
+
+### Database
+- **SQLite3** (Development and production database)
+
+### Deployment & Infrastructure
+- **Docker**: 5.0.2 (Containerization)
+- **Docker Compose**: 1.29.2 (Multi-container orchestration)
+- **Nginx** (Reverse proxy and static file serving)
+- **uWSGI**: >=2.0.19.1,<2.1 (WSGI server)
+- **Certbot** (SSL certificate management)
+
+### Build Tools
+- **Webpack**: 5.47.1 (Module bundler)
+- **Babel**: 7.14.8 (JavaScript transpiler)
+
+## Configuration
+
+### Environment Variables
+**Required:**
+- `SECRET_KEY`: Django secret key for cryptographic signing
+- `DEBUG`: Debug mode flag (1 for development, 0 for production)  
+- `ALLOWED_HOSTS`: Comma-separated list of allowed host/domain names
+
+**Optional:**
+- `SQLITE3_DB`: Database filename (defaults to db.sqlite3)
+- `APP_HOST`: Application host for proxy (defaults to app)
+- `APP_PORT`: Application port for proxy (defaults to 9000)
+
+### Email Configuration
+
+The application uses email for user account activation. Email behavior differs between development and production:
+
+**Development Mode (DEBUG=1):**
+- Uses `django.core.mail.backends.console.EmailBackend`
+- Emails are printed to console/Docker logs instead of being sent
+- No SMTP credentials required
+- View activation emails with: `docker-compose logs app | grep -A 20 "activation"`
+
+**Production Mode (DEBUG=0):**
+- Uses `django.core.mail.backends.smtp.EmailBackend`
+- `EMAIL_HOST`: smtp.gmail.com
+- `EMAIL_HOST_USER`: Environment variable or defaults to info.chunkitapp@gmail.com
+- `EMAIL_HOST_PASSWORD`: **Required** - Set via environment variable
+- `EMAIL_PORT`: 587
+- `EMAIL_USE_TLS`: True
+
+**Production Email Setup:**
+```bash
+# Set these environment variables for production
+export EMAIL_HOST_USER=your-email@gmail.com
+export EMAIL_HOST_PASSWORD=your-app-password  # Use Gmail App Password, not regular password
+```
+
+### Feature Flags
+- **ImitationTask**: Boolean field in experiment configuration
+- **UseQuestions**: Boolean field to enable comprehension questions
+- **UseProlific**: Boolean field to enable Prolific integration
[…]
