+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing Guidelines
+
+## Branch Naming
+- `feature/experiment-feature-name`
+- `bugfix/issue-description`
+- `hotfix/critical-issue`
+
+## Code Style
+- Follow Django coding standards
+- Use React/JavaScript ES6+ conventions
+- Maintain consistent indentation (4 spaces for Python, 2 for JavaScript)
+
+## Pull Request Process
+1. Fork the repository
+2. Create feature branch
+3. Make changes with appropriate tests
+4. Submit pull request with detailed description
+
+## Contact
+- Technical issues: GitHub issues
+- Research questions: alena.konina@helsinki.fi or alena.konina@gmail.com
