+++ docs/troubleshooting.md
@@ -1,0 +1,194 @@
+# Troubleshooting Guide
+
+## Common Issues & Solutions
+
+### Windows Users: Line Ending Issues
+
+If you see errors like `exec /scripts/run.sh: no such file or directory` or `/run.sh: not found`, this is caused by Windows line endings (CRLF) in shell scripts. Here's how to fix it:
+
+**Step 1: Fix Git Configuration**
+```bash
+# Configure Git to not convert line endings on Windows
+git config --global core.autocrlf input
+```
+
+**Step 2: Re-clone or Reset Repository**
+```bash
+# Option A: Fresh clone (recommended)
+cd ..
+rm -rf ChunkitApp2.0
+git clone https://github.com/annthehuman/ChunkitApp2.0.git
+cd ChunkitApp2.0
+
+# Option B: Reset existing repository
+git rm --cached -r .
+git reset --hard
+```
+
+**Step 3: Convert Existing Script Line Endings** (if you prefer not to re-clone)
+```bash
+# On Windows with Git Bash or WSL:
+dos2unix scripts/run.sh proxy/run.sh proxy/certbot/certify-init.sh entrypoint.sh
+
+# Or manually with PowerShell:
+(Get-Content scripts/run.sh -Raw) -replace "`r`n","`n" | Set-Content scripts/run.sh -NoNewline
+(Get-Content proxy/run.sh -Raw) -replace "`r`n","`n" | Set-Content proxy/run.sh -NoNewline
+```
+
+**Step 4: Rebuild Docker Images**
+```bash
+docker-compose down
+docker-compose build --no-cache
+docker-compose up
+```
+
+**Why This Happens:**
+Windows Git automatically converts line endings:
+- **LF** (`\n`) = Unix/Linux line ending
+- **CRLF** (`\r\n`) = Windows line ending
+
+When shell scripts have CRLF endings, Linux sees the shebang as `#!/bin/sh\r` (with an invisible `\r` character), which doesn't exist, causing the "no such file or directory" error.
+
+The `.gitattributes` file now forces all `.sh` files to always use LF endings, even on Windows.
+
+### Missing .env File
+
+If you see warnings about missing environment variables, create a `.env` file:
+
+```bash
+# Create .env file in project root
+cat > .env << 'EOF'
+SECRET_KEY=your-secret-key-here
+DEBUG=1
+ALLOWED_HOSTS=127.0.0.1,localhost
+EOF
+```
+
+Generate a secret key at: https://djecrety.ir/
+
+### Permission Issues (Linux/Mac)
+
+If you encounter permission errors:
+```bash
+chmod +x scripts/run.sh
+chmod +x proxy/run.sh
+chmod +x entrypoint.sh
+chmod +x proxy/certbot/certify-init.sh
+```
+
+### Database Permissions
+
+```bash
+# Fix SQLite permissions in Docker
+chmod 777 /vol/web/db.sqlite3
+chown app:app /vol/web/db.sqlite3
+```
+
+### Media File Upload Issues
+
+```bash
+# Ensure media directory permissions
+chmod -R 777 /vol/web/media
+mkdir -p /vol/web/media/Practice
+mkdir -p /vol/web/media/Experement
+```
+
+### Audio File Extraction Problems
+
+- Ensure ZIP files don't contain nested folders beyond one level
+- Audio files should be directly in ZIP or one folder deep
+- Avoid system files like `__MACOSX` folders
+
+### CORS Issues
+
+- Verify `CORS_ORIGIN_WHITELIST` includes your frontend URL
+- Check `ALLOWED_HOSTS` includes your domain
+
+### Email & Signup Issues
+
+```bash
+# In development (DEBUG=1): Activation emails print to Docker logs
[…]
