### 2026-01-29 #8486 @wmvanvliet
With the hugely improved installation instructions, I finally got Chunkit to run, and account creation now works nicely.

Unfortunately, after playing around a bit with the program, I am of the opinion it's not mature enough yet, as even basic functionality is broken. I was unable to get the program to say anything other than "Where is no data to display" no matter what kind of zip file with .wav files or transcript table I tried. Every radiobutton I have encountered along the way was also broken. The bugs are too numerous for me to move forward. No doubt these things are caused by small things that are easy to fix, but it's clear that nobody ever tried running ChunkIt in a place other than where it was created. Until ChunkIt has multiple users and thus actually runs at multiple locations, I fear it doesn't yet meet the bar for JOSS.

I'm sorry, I know you've put a lot of effort into this project, and I really wish for it to succeed, but it really needs a lot more work.
