### 2026-02-01 #8486 @akonina
@wmvanvliet thank you for your kind comments. it's disappointing for sure, but your delivery made it a bit better, thank you.

@samhforbes what could my next steps be? is it okay for me to resubmit when I feel like the app is ready?
