### 2026-02-05 #8486 @samhforbes
Hi @wmvanvliet  thanks for your review that's really helpful.

@akonina Yes that's right. I'll close this review issue but then when you feel it's ready you are welcome to resubmit. Please note that in the mean time our scope documentation has updated and any new submission will need to meet those updated scope requirements
